Support Input Transformations
Simulation and learner now support input transformations.
We changed the way challenges are stored to achieve this;
the list of challenges now contains N challenges, each
challenge modelled as an array of k individual challenges,
one for each Arbiter chain.

Unit tests have been changed to test according to the new
model.

The only shipped transformation as of now is the 'ATF'
transformation that models the 'Arbiter Threshold
Function', that is, the monomials that arise when modeling
an Arbiter PUF as a polynomial threshold function.

sim_learn.py now outputs a comprehensive, aggregated
result instead of individual results.

Some run time optimization included.