Implemented Polynomial Inputtransformation
The goal of the polynomial inputtransformation is to derive pseudorandom
inputs for each of the k Arbiter Chains of the PUF from one challenge c.
For this, we interpret c as a polynomial from the finite field
GF(2^len(c)) = F2/f*F2, where f is a irreducible polynomial of degree
len(c).
To derive k challenges we compute the list
[c^2 mod f, c^3 mod f, ..., c^(k+1) mod f]
which gives us polynomials from F2/f*F2. Their coefficients are interpreted
as input bits for the Arbiter Chains.

For now the irreducible polynomial f is hard coded and of degree 64.

The file pypuf/tooly.py was expended by functions to transfrom the
challenges from -1,1 notation to 0,1 notation and back.