Fixes issue  #31 and #88
This commit introduces a better bias parameter.
The bias parameter could be a float or a two
dimensional array of float with shape (k, 1).
If bias is a float then each PUF gets the
same bias. If bias is a two dimensional array
then each PUF gets its own bias. If the bias
is None then no bias is used (by default).
The bias is added after input transformation
to ensure that it is used. This commit also
fixes a bug which occured in
ExperimentMajorityVoteFindVotes (see #88).